blend cli integration with new CLI setup. create config methods for creating main run configs, blend config, and agg configs, which can then be used as inputs to cli modules
