+++ nsrdb/config/var_descriptions.csv
@@ -1,6 +1,6 @@
 ﻿var,description
 air_temperature,"Air temperature at 2 meters above the surface."
-alpha,"Alpha parameter specific to the dataset, representing a coefficient or scaling factor."
+alpha,"Angstrom wavelength exponent, ideally obtained by linear regression of all available spectral AODs between 380 and 1020 nm."
 aod,"Aerosol Optical Depth (AOD) measures the attenuation of solar radiation due to aerosol particles in the atmosphere, providing insight into air quality and visibility."
 asymmetry,"Asymmetry parameter specific to the dataset, often related to the directional properties of scattered radiation or aerosol properties."
 cld_opd_dcomp,"Cloud optical depth decomposed specific to the dataset, indicating the optical thickness of clouds and how it varies across different components."
@@ -17,11 +17,11 @@ dni,"Direct normal irradiance is the amount of solar radiation received per unit
 ghi,"Global horizontal irradiance is the total amount of solar radiation received per unit area on a horizontal surface."
 fill_flag,"Fill flag indicating data quality or presence, often used to identify missing or interpolated data points."
 cloud_fill_flag,"Cloud fill flag indicating data quality or presence, specific to cloud-related data points."
-ozone,"Ozone concentration in the atmosphere, typically measured in Dobson units (DU) or parts per billion (ppb)."
-relative_humidity,"Relative humidity is the ratio of the amount of water vapor present in the air to the maximum amount that the air could hold at the given temperature, expressed as a percentage."
-solar_zenith_angle,"Solar zenith angle is the angle between the zenith (straight overhead point) and the position of the sun, measured from the observer's point on the Earth's surface."
-specific_humidity,"Specific humidity is the mass of water vapor present per unit mass of air, typically measured in grams of water vapor per kilogram of dry air."
-ssa,"Single Scattering Albedo (SSA) measures the ratio of the scattering coefficient to the extinction coefficient, providing information about the scattering properties of aerosols or particles."
+ozone,"Reduced ozone vertical pathlength."
+relative_humidity,"Ratio of the amount of water vapor present in the air to the maximum amount that the air could hold at the given temperature, expressed as a percentage."
+solar_zenith_angle,"Angle between the zenith (straight overhead point) and the position of the sun, measured from the observer's point on the Earth's surface."
+specific_humidity,"Mass of water vapor present per unit mass of air, typically measured in grams of water vapor per kilogram of dry air."
+ssa,"Aerosol single-scattering albedo at a representative wavelength of about 700 nm. Measures the ratio of the scattering coefficient to the extinction coefficient."
 surface_albedo,"Surface albedo is the fraction of solar energy reflected by the Earth's surface, often varying based on surface type (e.g., snow-covered vs. vegetated)."
 surface_pressure,"Surface atmospheric pressure is the force per unit area exerted by the weight of the atmosphere above a specific location on the Earth's surface."
 total_precipitable_water,"Total precipitable water is the depth of liquid water that would result from precipitating all the water vapor in a vertical column above a given location, often measured in millimeters or centimeters."
