+++ tests/postproc/test_aggregation.py
